330 Moves logic to ContentContext and useContentRef
