+++ packages/orchestrator-ui-components/src/components/WfoPageTemplate/WfoPageTemplate/index.ts
@@ -1,1 +1,2 @@
+export * from './ContentContext';
 export * from './WfoPageTemplate';
