+++ packages/orchestrator-ui-components/src/hooks/useContentRef.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { ContentContext } from '../components/WfoPageTemplate';
+
+export const useContentRef = () => ({
+    contentRef: useContext(ContentContext)?.contentRef,
+});
